Implementacao do projeto Inbound

+++ src/Services/InbounOrder/Inbound.Domain/OrderItem.cs
@@ -4,7 +4,7 @@ namespace Inbound.Domain
 {
     public class OrderItem : Entity
     {
-        public OrderItem(Guid documentId, string productCode, string typePackage, decimal quantityPackage, decimal quantity)
+        public OrderItem(Guid documentId, string productCode, string typePackage, int quantityPackage, decimal quantity)
         {
             DocumentId = documentId;
             ProductCode = productCode;
@@ -16,7 +16,7 @@ namespace Inbound.Domain
         public Guid DocumentId { get; set; }
         public string ProductCode { get; private set; }
         public string TypePackage { get; private set; }
-        public decimal QuantityPackage { get; private set; }
+        public int QuantityPackage { get; private set; }
         public decimal Quantity { get; private set; }
     }
 }
